moved upload_blob to blob.py, upload-batch calls upload_blob
