+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/custom.py
@@ -13,7 +13,6 @@ from azure.cli.core.profiles import get_sdk, supported_api_version, ResourceType
 
 from azure.cli.command_modules.storage._factory import storage_client_factory
 from azure.cli.command_modules.storage.util import guess_content_type
-from azure.cli.core.application import APPLICATION
 from .sdkutil import get_table_data_type
 
 Logging, Metrics, CorsRule, AccessPolicy, RetentionPolicy = \
@@ -26,15 +25,6 @@ BlockBlobService, BaseBlobService, FileService, FileProperties, DirectoryPropert
             'queue#QueueService')
 
 TableService = get_table_data_type('table', 'TableService')
-
-
-def _update_progress(current, total):
-    HOOK = APPLICATION.get_progress_controller(True)
-
-    if total:
-        HOOK.add(message='Alive', value=current, total_val=total)
-        if total == current:
-            HOOK.end()
 
 
 # CUSTOM METHODS
@@ -210,83 +200,6 @@ def show_storage_account_connection_string(
     connection_string = '{}{}'.format(connection_string,
                                       ';TableEndpoint={}'.format(table_endpoint) if table_endpoint else '')
     return {'connectionString': connection_string}
-
-
-@transfer_doc(BlockBlobService.create_blob_from_path)
-def upload_blob(client, container_name, blob_name, file_path, blob_type=None, content_settings=None, metadata=None,
-                validate_content=False, maxsize_condition=None, max_connections=2, lease_id=None, tier=None,
-                if_modified_since=None, if_unmodified_since=None, if_match=None, if_none_match=None, timeout=None):
-    """Upload a blob to a container."""
-
-    settings_class = get_sdk(ResourceType.DATA_STORAGE, 'blob.models#ContentSettings')
-    content_settings = guess_content_type(file_path, content_settings, settings_class)
-
-    def upload_append_blob():
-        if not client.exists(container_name, blob_name):
-            client.create_blob(
-                container_name=container_name,
-                blob_name=blob_name,
-                content_settings=content_settings,
-                metadata=metadata,
-                lease_id=lease_id,
-                if_modified_since=if_modified_since,
-                if_match=if_match,
-                if_none_match=if_none_match,
-                timeout=timeout)
-
-        append_blob_args = {
-            'container_name': container_name,
-            'blob_name': blob_name,
-            'file_path': file_path,
-            'progress_callback': _update_progress,
-            'maxsize_condition': maxsize_condition,
-            'lease_id': lease_id,
-            'timeout': timeout
-        }
-
-        if supported_api_version(ResourceType.DATA_STORAGE, min_api='2016-05-31'):
-            append_blob_args['validate_content'] = validate_content
-
-        return client.append_blob_from_path(**append_blob_args)
-
-    def upload_block_blob():
-        import os
-
-        # increase the block size to 100MB when the file is larger than 200GB
-        if os.path.isfile(file_path) and os.stat(file_path).st_size > 200 * 1024 * 1024 * 1024:
-            client.MAX_BLOCK_SIZE = 100 * 1024 * 1024
-            client.MAX_SINGLE_PUT_SIZE = 256 * 1024 * 1024
-
-        create_blob_args = {
-            'container_name': container_name,
-            'blob_name': blob_name,
-            'file_path': file_path,
-            'progress_callback': _update_progress,
-            'content_settings': content_settings,
-            'metadata': metadata,
-            'max_connections': max_connections,
-            'lease_id': lease_id,
-            'if_modified_since': if_modified_since,
-            'if_unmodified_since': if_unmodified_since,
-            'if_match': if_match,
-            'if_none_match': if_none_match,
-            'timeout': timeout
-        }
-
-        if supported_api_version(ResourceType.DATA_STORAGE, min_api='2017-04-17') and tier:
-            create_blob_args['premium_page_blob_tier'] = tier
-
-        if supported_api_version(ResourceType.DATA_STORAGE, min_api='2016-05-31'):
-            create_blob_args['validate_content'] = validate_content
-
-        return client.create_blob_from_path(**create_blob_args)
-
-    type_func = {
-        'append': upload_append_blob,
-        'block': upload_block_blob,
-        'page': upload_block_blob  # same implementation
-    }
-    return type_func[blob_type]()
 
 
 def set_blob_tier(client, container_name, blob_name, tier, blob_type='block', timeout=None):
